Use legacy config file in training.py
