+++ utils/config.py
@@ -1,8 +1,6 @@
 
 import configparser
-import pathlib
 import os
-import shutil
 
 FUNCTIONS_SECTION = 'FUNCTIONS'
 FUNCTIONS_KEY = 'FUNCTIONS_KEY'
@@ -131,18 +129,6 @@ class Config():
         parser.read(config_path)
         return Config.read_config_with_parsed_config(parser)
 
-    @staticmethod
-    def initialize_training_location(config):
-        file_tree = pathlib.Path(os.path.expanduser(
-            config.get("training_location"))
-        )
-
-        if file_tree.exists():
-            print("Removing existing tag data directory: " + str(file_tree))
-
-            shutil.rmtree(str(file_tree), ignore_errors=True)
-
-        return pathlib.Path(file_tree)
 
 class MissingConfigException(Exception):
     pass
